In case of null links check for end-of-stream (#1138)
## Description
<!-- Provide a brief summary of the changes made and the issue they aim
to address.-->
NO_CHANGELOG=true

In case of no/null links, check for end-of-stream. This will prevent
driver making fetch calls to the server for closed statements.

## Testing
<!-- Describe how the changes have been tested-->
- manual testing
- unit tests

## Additional Notes to the Reviewer
<!-- Share any additional context or insights that may help the reviewer
understand the changes better. This could include challenges faced,
limitations, or compromises made during the development process.
Also, mention any areas of the code that you would like the reviewer to
focus on specifically. -->